feat: service configuration and history search processing

+++ src/main/java/com/group4/swissrouteapi/dtos/responses/history/HistoryResponse.java
@@ -16,6 +16,6 @@ public class HistoryResponse {
   List<History> history;
   Integer page;
   Integer size;
-  Integer totalElements;
+  Long totalElements;
   Integer totalPages;
 }
